+++ vector/src/main/java/im/vector/riotx/features/roomprofile/banned/RoomBannedViewEvents.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2020 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package im.vector.riotx.features.roomprofile.banned
+
+import im.vector.matrix.android.api.session.room.model.RoomMemberSummary
+import im.vector.riotx.core.platform.VectorViewEvents
+
+sealed class RoomBannedViewEvents : VectorViewEvents {
+    data class ShowBannedInfo(val bannedByUserId: String, val banReason: String, val roomMemberSummary: RoomMemberSummary) : RoomBannedViewEvents()
+    data class ToastError(val info: String) : RoomBannedViewEvents()
+}
